Moving nested test class about BasicAuthenticationManager
