+++ caching-service/src/test/java/org/zowe/apiml/caching/config/SpringSecurityConfigTest.java
@@ -17,40 +17,26 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.ReactiveAuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
 import org.zowe.apiml.caching.CachingServiceApplication;
-import org.zowe.apiml.security.common.auth.BasicAuthenticationManager;
 import org.zowe.apiml.util.config.SslContext;
 import org.zowe.apiml.util.config.SslContextConfigurer;
-import reactor.test.StepVerifier;
 
 import java.util.Base64;
-import java.util.Collections;
 
 import static io.restassured.RestAssured.given;
-import static org.junit.jupiter.api.Assertions.*;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class SpringSecurityConfigTest {
 
     private static final String USER = "user";
     private static final String PASSWORD = "password";
-    private static final String ROLE = "CACHING_SERVICE";
 
-    private static final AbstractAuthenticationToken VALID_AUTHENTICATION = new UsernamePasswordAuthenticationToken(USER, PASSWORD.toCharArray(), Collections.singleton(new SimpleGrantedAuthority("CACHING_SERVICE")));
     private static final String VALID_BASIC_AUTH = "Basic " + Base64.getEncoder().encodeToString((USER + ":" + PASSWORD).getBytes());
     private static final String INVALID_BASIC_AUTH = "Basic " + Base64.getEncoder().encodeToString((USER + ":invalidPassword").getBytes());
     private static final String X_CS_SERVICE_ID = "X-CS-Service-ID";
@@ -381,83 +367,6 @@ public class SpringSecurityConfigTest {
                     .log().ifValidationFails()
                     .statusCode(HttpStatus.OK.value());
             }
-        }
-
-    }
-
-    @Nested
-    class BasicAuthentication {
-
-        @Nested
-        class MissingCredentials {
-
-            @ParameterizedTest
-            @CsvSource({
-                ",,",
-                "user,,",
-                ",password,"
-            })
-            void givenNoCompleteCredentials_whenAuthorize_thenThrowException(String user, String password) {
-                var authenticationManager = new BasicAuthenticationManager(user, password == null ? null : password.toCharArray(), ROLE);
-                StepVerifier.create(authenticationManager.authenticate(VALID_AUTHENTICATION))
-                    .expectError(BadCredentialsException.class)
-                    .verify();
-            }
-
-        }
-
-        @Nested
-        class ValidCredentials {
-
-            private ReactiveAuthenticationManager basicAuthenticationManager = new BasicAuthenticationManager(USER, PASSWORD.toCharArray(), ROLE);
-
-            @Test
-            void givenValidCredentials_whenAuthenticate_thenSuccess() {
-                StepVerifier.create(basicAuthenticationManager.authenticate(VALID_AUTHENTICATION))
-                    .assertNext(authentication -> {
-                        assertNotSame(VALID_AUTHENTICATION, authentication);
-                        assertTrue(authentication.isAuthenticated());
-                        assertEquals(USER, authentication.getName());
-                        String credentials = authentication.getCredentials() instanceof char[] chars ? new String(chars) : String.valueOf(authentication.getCredentials());
-                        assertEquals(PASSWORD, credentials);
-                    })
-                    .verifyComplete();
-            }
-
-            @Test
-            void givenValidCredentialsInAnotherForm_whenAuthenticate_thenSuccess() {
-                var user = new Object() {
-                    @Override
-                    public String toString() {
-                        return USER;
-                    }
-                };
-                Authentication validAuthentication = new UsernamePasswordAuthenticationToken(user, PASSWORD, Collections.singleton(new SimpleGrantedAuthority(ROLE)));
-                StepVerifier.create(basicAuthenticationManager.authenticate(validAuthentication))
-                    .assertNext(authentication -> {
-                        assertNotSame(validAuthentication, authentication);
-                        assertTrue(authentication.isAuthenticated());
-                        assertEquals(USER, authentication.getName());
-                        assertTrue(authentication.getCredentials() instanceof char[]);
-                        assertEquals(PASSWORD, new String((char[]) authentication.getCredentials()));
-                    })
-                    .verifyComplete();
-            }
-
-            @ParameterizedTest
-            @CsvSource({
-                ",,",
-                "user,,",
-                ",password,",
-                "attacker,attempt"
-            })
-            void givenInvalidCredentials_whenAuthenticate_thenThrowException(String user, String password) {
-                var authentication = new UsernamePasswordAuthenticationToken(user, password == null ? null : password.toCharArray(), Collections.singleton(new SimpleGrantedAuthority(ROLE)));
-                StepVerifier.create(basicAuthenticationManager.authenticate(authentication))
-                    .expectError(BadCredentialsException.class)
-                    .verify();
-            }
-
         }
 
     }
